pyproject.toml
[tool.poetry]
name = "pineconeexperiment"
version = "0.1.0"
description = "An experiment with the pinecone database. Current goal is to do a simple similarity search with imagenet and the imagenette2 database"
authors = ["FRRitsma <folkert.ritsma@xomnia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pinecone-client = "^2.2.1"
torch = "^2.0.1"
torchvision = "^0.15.2"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
structlog = "^23.1.0"
requests = "^2.31.0"
types-requests = "^2.31.0.1"
mypy = "^1.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pre-commit = "^3.3.1"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.0"
pytest-mock = "^3.10.0"
poethepoet = "^0.10.0"

[tool.poe.tasks]
force-cuda = "pip install torch==2.0.1+cu117 -f https://download.pytorch.org/whl/cu117/torch_stable.html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
markers = [
    "needs_internet: Denotes that a test requires internet to be succeed (deselect with '-m \"not needs_internet\"')",
]

[tool.poetry.scripts]
post-install = "poe force-cuda"
